telegram-bot/requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
qrcode[pil]==7.4.2
pyyaml==6.0.2
